a couple of formatting and guideline compliance changes
